+++ slim/base/app.py
@@ -111,6 +111,7 @@ class Application:
         for vi in self.route.views:
             cls = vi.view_cls
             if issubclass(cls, AbstractSQLView):
+                '''
                 def get_exists_view_full_name():
                     view_cls = self.tables[cls.table_name]
                     return '%s.%s' % (view_cls.__module__, view_cls.__name__)
@@ -118,7 +119,6 @@ class Application:
                 def get_view_full_name():
                     return '%s.%s' % (cls.__module__, cls.__name__)
 
-                '''
                 if cls.table_name in self.tables:
                     logger.error("Binding %r failed, the table %r is already binded to %r." % (get_view_full_name(), cls.table_name, get_exists_view_full_name()))
                     exit(-1)